requirements.txt
# Modern Turkish POS Tagger Requirements
# Updated for Python 3.9+ with modern ML libraries

# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.12.0

# Traditional NLP
nltk>=3.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Utilities
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.65.0

# Web Service
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Optional: For advanced features
# tensorflow>=2.13.0  # Alternative to PyTorch
# spaCy>=3.6.0        # Alternative NLP library
# fastapi>=0.100.0    # Alternative to Flask